+++ include/BALL/MOLMEC/AMBER/amberTorsion.h
@@ -1,4 +1,4 @@
-// $Id: amberTorsion.h,v 1.6 2000/02/14 22:43:56 oliver Exp $
+// $Id: amberTorsion.h,v 1.7 2000/03/26 12:45:42 oliver Exp $
 // Molecular Mechanics: Amber force field, bond stretch component
 
 #ifndef BALL_MOLMEC_AMBER_AMBERTORSION_H
@@ -103,7 +103,7 @@ namespace BALL
 
 		/**	Calculates and returns the component's energy.
 		*/
-		virtual float updateEnergy();
+		virtual double updateEnergy();
 
 		/**	Calculates and returns the component's forces.
 		*/
